+++ Sources/Sentry/Public/SentrySDK.h
@@ -308,6 +308,15 @@ SENTRY_NO_INIT
 + (void)crash;
 
 /**
+ * Waits synchronously for the SDK to flush out all queued and cached items for up to the specified
+ * timeout in seconds. If there is no internet connection, the function returns immediately. The SDK
+ * doesn't dispose the client or the hub.
+ *
+ * @param timeout The time to wait for the SDK to complete the flush.
+ */
++ (void)flush:(NSTimeInterval)timeout NS_SWIFT_NAME(flush(timeout:));
+
+/**
  * Closes the SDK and uninstalls all the integrations.
  */
 + (void)close;