+++ Sources/Sentry/SentryClient.m
@@ -89,7 +89,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
                             andCurrentDateProvider:[SentryDefaultCurrentDateProvider sharedInstance]
                                              error:&error];
     if (nil != error) {
-        [SentryLog logWithMessage:error.localizedDescription andLevel:kSentryLevelError];
+        SENTRY_LOG_ERROR(@"%@", error.localizedDescription);
         return nil;
     }
 
@@ -382,7 +382,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
         if (nil == session.releaseName || [session.releaseName length] == 0) {
             SentryTraceContext *traceContext = [self getTraceStateWithEvent:event withScope:scope];
 
-            [SentryLog logWithMessage:DropSessionLogMessage andLevel:kSentryLevelDebug];
+            SENTRY_LOG_DEBUG(DropSessionLogMessage);
 
             [self.transportAdapter sendEvent:event
                                 traceContext:traceContext
@@ -402,7 +402,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
 - (void)captureSession:(SentrySession *)session
 {
     if (nil == session.releaseName || [session.releaseName length] == 0) {
-        [SentryLog logWithMessage:DropSessionLogMessage andLevel:kSentryLevelDebug];
+        SENTRY_LOG_DEBUG(DropSessionLogMessage);
         return;
     }
 
@@ -434,8 +434,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
     }
 
     if ([SentryId.empty isEqual:userFeedback.eventId]) {
-        [SentryLog logWithMessage:@"Capturing UserFeedback with an empty event id. Won't send it."
-                         andLevel:kSentryLevelDebug];
+        SENTRY_LOG_DEBUG(@"Capturing UserFeedback with an empty event id. Won't send it.");
         return;
     }
 
@@ -462,12 +461,21 @@ NSString *const kSentryDefaultEnvironment = @"production";
                   isCrashEvent:NO];
 }
 
+- (void)flush:(NSTimeInterval)timeout
+{
+    [self.transportAdapter flush:timeout];
+}
+
 - (SentryEvent *_Nullable)prepareEvent:(SentryEvent *)event
                              withScope:(SentryScope *)scope
                 alwaysAttachStacktrace:(BOOL)alwaysAttachStacktrace
                           isCrashEvent:(BOOL)isCrashEvent
 {
     NSParameterAssert(event);
+    if (event == nil) {
+        return nil;
+    }
+
     if ([self isDisabled]) {
         [self logDisabledMessage];
         return nil;
@@ -478,8 +486,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
 
     // Transactions have their own sampleRate
     if (eventIsNotATransaction && [self isSampled:self.options.sampleRate]) {
-        [SentryLog logWithMessage:@"Event got sampled, will not send the event"
-                         andLevel:kSentryLevelDebug];
+        SENTRY_LOG_DEBUG(@"Event got sampled, will not send the event");
         [self recordLostEvent:kSentryDataCategoryError reason:kSentryDiscardReasonSampleRate];
         return nil;
     }
@@ -600,8 +607,7 @@ NSString *const kSentryDefaultEnvironment = @"production";
 
 - (void)logDisabledMessage
 {
-    [SentryLog logWithMessage:@"SDK disabled or no DSN set. Won't do anyting."
-                     andLevel:kSentryLevelDebug];
+    SENTRY_LOG_DEBUG(@"SDK disabled or no DSN set. Won't do anyting.");
 }
 
 - (SentryEvent *_Nullable)callEventProcessors:(SentryEvent *)event
@@ -611,9 +617,8 @@ NSString *const kSentryDefaultEnvironment = @"production";
     for (SentryEventProcessor processor in SentryGlobalEventProcessor.shared.processors) {
         newEvent = processor(newEvent);
         if (nil == newEvent) {
-            [SentryLog logWithMessage:@"SentryScope callEventProcessors: An event "
-                                      @"processor decided to remove this event."
-                             andLevel:kSentryLevelDebug];
+            SENTRY_LOG_DEBUG(@"SentryScope callEventProcessors: An event processor decided to "
+                             @"remove this event.");
             break;
         }
     }