+++ Tests/SentryTests/Integrations/SentryBaseIntegrationTests.swift
@@ -9,20 +9,24 @@ class MyTestIntegration: SentryBaseIntegration {
 
 class SentryBaseIntegrationTests: XCTestCase {
     var logOutput: TestLogOutput!
+    var oldDebug: Bool!
+    var oldLevel: SentryLevel!
+    var oldOutput: SentryLogOutput!
 
     override func setUp() {
         super.setUp()
+        oldDebug = SentryLog.isDebug()
+        oldLevel = SentryLog.diagnosticLevel()
+        oldOutput = SentryLog.logOutput()
         SentryLog.configure(true, diagnosticLevel: SentryLevel.debug)
-
         logOutput = TestLogOutput()
         SentryLog.setLogOutput(logOutput)
     }
 
     override func tearDown() {
         super.tearDown()
-        // Set back to default
-        SentryLog.configure(true, diagnosticLevel: SentryLevel.error)
-        SentryLog.setLogOutput(nil)
+        SentryLog.configure(oldDebug, diagnosticLevel: oldLevel)
+        SentryLog.setLogOutput(oldOutput)
     }
 
     func testIntegrationName() {