+++ Tests/SentryTests/Transaction/SentrySpanTests.swift
@@ -284,7 +284,7 @@ class SentrySpanTests: XCTestCase {
     func testSpanWithoutTracer_StartChild_ReturnsNoOpSpan() {
         // Span has a weak reference to tracer. If we don't keep a reference
         // to the tracer ARC will deallocate the tracer.
-        let sutGenerator : () -> Span = {
+        let sutGenerator: () -> Span = {
             let tracer = SentryTracer()
             return SentrySpan(tracer: tracer, context: SpanContext(operation: ""))
         }